Extract client
